int division

+++ net/views/onthesky.py
@@ -114,7 +114,7 @@ def plot_wcs_outline(wcsfn, plotfn, W=256, H=256, width=36, zoom=True,
 
 def plot_aitoff_wcs_outline(wcsfn, plotfn, W=256, zoom=True):
     #anutil.log_init(3)
-    H = W/2
+    H = int(W/2)
     # Create Hammer-Aitoff WCS of the appropriate size.
     wcs = anutil.anwcs_create_allsky_hammer_aitoff(0., 0., W, H)
